# [IMPROVED] Route performance for larger messages