Similar improvement that we got in the client's parser.

The PR includes benchmarks for routes. Here would be the results with/without the changes.
```
Original:
Benchmark_RoutePubSub_NoPayload-8	 3000000	       590 ns/op
Benchmark_RoutePubSub_1K-8       	  500000	      3119 ns/op
Benchmark_RoutePubSub_100K-8     	   10000	    128411 ns/op

Change in parser:
Benchmark_RoutePubSub_NoPayload-8	 3000000	       584 ns/op
Benchmark_RoutePubSub_1K-8       	 1000000	      1181 ns/op (2.64x faster)
Benchmark_RoutePubSub_100K-8     	   20000	     76809 ns/op (1.67x faster)
```

/cc @derekcollison